split integer_arithmetic.py
